+++ component-classes/index.d.ts
@@ -516,6 +516,11 @@ export namespace backwardsCompatibleClasses {
     export let removedRadioTokens: string;
     export let removedSelectTokens: string;
     export let removedSliderTokens: string;
+    export let removedStepindicatorTokens: string;
+    export let removedSwitchTokens: string;
+    export let removedTabsTokens: string;
+    export let removedToastTokens: string;
+    export let removedTooltipTokens: string;
 }
 export namespace pagination {
     export let currentPage: string;